+++ HISTORY.rst
@@ -3,6 +3,11 @@
 Release History
 ---------------
 
+Unreleased
+++++++++++
+
+- fix: #608 Name of placeholder in layout gets reset to placeholder basename when adding slide
+
 1.0.2 (2024-08-07)
 ++++++++++++++++++